// src/shared/generated/ExtensionSnapshotRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ExtensionPackageRecord } from "./ExtensionPackageRecord";
import type { ExtensionRoleRecord } from "./ExtensionRoleRecord";

export type ExtensionSnapshotRecord = { revision: number, installed: Array<ExtensionPackageRecord>, roles: Array<ExtensionRoleRecord>, };